+++ src/modules/statistic/statistic-change-schedule/state/statistic-change-schedule.effects.ts
@@ -74,7 +74,7 @@ export class StatisticChangeScheduleEffects {
     };
 
     return this.statisticService
-      .getChangeSchedule(
+      .getDepartment(
         departmentId,
         UrlHelper.queryFilter(
           statisticData,