+++ src/services/statistic.service.ts
@@ -7,6 +7,7 @@ import {
   ChangeScheduleStatistic,
   ResponseModel,
 } from '@shared/models';
+import { TuiDayRange } from '@taiga-ui/cdk';
 import { Observable } from 'rxjs';
 import { BaseDataService } from './core/base-data.service';
 
@@ -18,7 +19,7 @@ export class StatisticService extends BaseDataService {
     super();
   }
 
-  public getChangeSchedule(
+  public getDepartment(
     departmentId: string,
     params: QueryFilterResult<ChangeScheduleStatistic, string>
   ): Observable<ResponseModel<ChangeSchedule[]>> {
@@ -27,8 +28,21 @@ export class StatisticService extends BaseDataService {
     ) as unknown as ChangeScheduleStatisticDta;
 
     return this.http.get<ResponseModel<ChangeSchedule[]>>(
-      this.url + `departments/${departmentId}/fixed-schedules?`,
+      this.url + `departments/${departmentId}/fixed-schedules`,
       { params: { ...parseParams } }
+    );
+  }
+
+  public getPersonal(
+    range: TuiDayRange,
+    teacherId: string
+  ): Observable<ResponseModel<ChangeSchedule[]>> {
+    const from = range.from.getFormattedDay('YMD', '-');
+    const to = range.to.getFormattedDay('YMD', '-');
+
+    return this.http.get<ResponseModel<ChangeSchedule[]>>(
+      this.url +
+        `teachers/${teacherId}/fixed-schedules?date=${from},${to}&old_date[sort]=asc&old_shift[sort]=asc&old_id_room[sort]=asc`
     );
   }
 }